feature: Implement movie listing and details screens (#1)
* feat: Add IconProps interface and update HeaderMenu and LogoutButton components

This commit adds the IconProps interface to define the props for SVG icons. It also updates the HeaderMenu and LogoutButton components to use the IconProps interface. This ensures consistent and reusable props for SVG icons throughout the application.

* feat: Update Home component to render a list of mocked MoovieCard images

The code changes in this commit update the Home component to implement lazy loading for the images displayed in the MoovieCard components. This improves the page load performance by only loading the images when they are visible on the screen. It enhances the user experience by reducing the initial load time and optimizing the memory usage.

* refactor: rename typo on component names and add env variables on gitIgnore

* feat: Add MoviesProvider to AppProvider

This commit adds the MoviesProvider to the AppProvider in the contexts/index.tsx file. By including the MoviesProvider, the application now has access to the movies context, which provides data and functions related to movies. This change allows for better organization and separation of concerns within the application's context architecture.

* feat: Implement movies context for movie listing in Home component

* feat: Add backgroundColor prop to IconProps interface and update ArrowBack and Lock icons

This commit adds the `backgroundColor` prop to the `IconProps` interface, allowing for the customization of the background color of SVG icons. It also updates the `ArrowBack` and `Lock` icons to utilize the `backgroundColor` prop, providing more flexibility in styling the icons. This change enhances the visual consistency and design options for the application's icons.

* feat: Update TabLayout component to use SafeAreaView and StatusBar

This commit updates the TabLayout component to use the SafeAreaView and StatusBar components from react-native-safe-area-context and react-native respectively. The SafeAreaView ensures that the content is displayed within the safe area boundaries of the device, while the StatusBar sets the background color and style of the status bar. This change improves the visual consistency and user experience of the TabLayout component.

* refactor: Update _layout.tsx to remove SafeAreaView

This commit removes the SafeAreaView component from the _layout.tsx file. The SafeAreaView was previously used to ensure that the content is displayed within the safe area boundaries of the device. However, it is no longer necessary as the Stack component from expo-router already handles safe area insets. This change simplifies the code and improves performance.